new_app/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.10.0
sentence-transformers>=2.2.2
chromadb>=0.5.0
python-pptx>=0.6.21
Pillow>=10.0.0
setuptools>=65.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pytesseract>=0.3.13
